+++ apps/client/src/app/pages/gathering-location/gathering-location/gathering-location.component.ts
@@ -143,6 +143,9 @@ export class GatheringLocationComponent {
     if (node.baits) {
       alarm.baits = node.baits;
     }
+    if (node.hookset) {
+      alarm.hookset = node.hookset;
+    }
     return this.alarmsFacade.applyFishEyes(alarm);
   }